+++ build_schema.py
@@ -2,34 +2,54 @@
 #coding:utf-8
 
 from pathlib import Path
-import sys
 import os
+import shutil
 import toml
-import master_type
+from repository_path import KanjiPath
+from cpp_source_generator import DataHppGenerator, RepositoryHppGenerator, RepositoryCppGenerator
+from master_type import MasterDataTypeInfo, MasterDataTypeManager
+from logging import getLogger, basicConfig, DEBUG, INFO
+logger = getLogger(__name__)
 
-class PathStr:
-    ROOT = '/Users/kazuaki/Documents/develop/youriko/KANJI/'
-    CLIENT = 'KANJI-client/'
-    DX4SIV3D = 'KANJI-client/Game/KANJI/dx4siv3d/'
-    ASSET = 'KANJI-client/Game/KANJI/KANJI-asset/'
-    SCHEMA = 'schema/'
-    TOOL = 'tool/'
+def output(file_path:Path, text:str):
+    os.makedirs(file_path.parent.absolute(), exist_ok=True)
+    logger.info('create file: %s' % file_path.relative_to(KanjiPath.absolute('md_header')))
+    with open(file_path,'w') as f:
+        f.write(text)
 
-class MasterDataActual:
-    def __init__(self, toml):
-        self.fields = toml
 
-if __name__ == '__main__':
-    path_schema = Path(PathStr.ROOT + PathStr.ASSET + PathStr.SCHEMA)
-    path_src = path_schema / 'master' / 'class'
-    path_dst = path_schema / 'master_header'
-    md_type_mgr = master_type.MasterDataTypeManager()
+def create_cpp_source(generator, type_info:MasterDataTypeInfo, file_path:Path):
+    full_text = generator.generate(type_info.data_type_name, type_info.fields)
+    output(file_path, full_text)
 
-    for path_toml in path_src.glob('**/*.toml'): # e.g. 'class/kanji/KanjiParam'
-        type_name = path_toml.stem # 'KanjiParam'
-        directories = path_toml.parent.relative_to(path_src) # kanji
-        type_info = md_type_mgr.at(type_name, str(directories))
-        file_dst = path_dst / directories
-        os.makedirs(file_dst, exist_ok=True)
-        os.chmod(file_dst, 0o755)
-        type_info.createHpp(file_dst)
+def create_cpp_sources_impl(type_info:MasterDataTypeInfo, path:Path, key:str):
+    indent = '    '
+    # MasterHoge.hpp
+    create_cpp_source(
+        DataHppGenerator(indent),
+        type_info,
+        path / 'class' / key / (type_info.data_type_name + '.hpp'))
+    # MasterHogeRepository.hpp
+    create_cpp_source(
+        RepositoryHppGenerator(indent),
+        type_info,
+        path / 'repository' / key / (type_info.data_type_name + 'Repository.hpp'))
+    # MasterHogeRepository.cpp
+    create_cpp_source(
+        RepositoryCppGenerator(indent),
+        type_info,
+        path / 'repository' / key / (type_info.data_type_name + 'Repository.cpp'))
+
+def create_cpp_sources(mgr:MasterDataTypeManager, path_dst:Path):
+    for key in mgr.dict_info:
+        for value in mgr.dict_info[key].values():
+            create_cpp_sources_impl(value, path_dst, key)
+
+if __name__ == "__main__":
+    basicConfig(level=INFO)
+    dest_dir = KanjiPath.absolute('md_header')
+    if os.path.isdir(dest_dir):
+        shutil.rmtree(dest_dir)
+    mgr = MasterDataTypeManager()
+    # mgr.create_cpp_source(dest_dir)
+    create_cpp_sources(mgr, dest_dir)
